Integrate pymunk physics engine
The current physical simulation is rather crude. It would be much nicer
to use a real physics engine. So let's!

Because of how I've architected things, with all physical simulation
being encapsulated by the physics subsystem, this was comically easy and
the game itself didn't even need to be modified. Hurrah!