Introduce thrusters
Previously the player was moved around by spookily updating the
velocity. Now, it has a "Thrusters" component which contains a number of
thrusters and can apply forces to the player in order to move in in
particular directions.

The idea is to have everything eventually use thrusters, but that will
require some more complex code. For now it's just the player since the
keyboard keys can map directly to the engines. (But this is very hard to
control so will need to change too.)