Add an enemy with turrets on it.
They don't really work yet! To see them in action, change the enemy type
in WaveSpawner to "destroyer.txt".

Problems: They try to shoot through themselves. Could either make
shooting smarter and avoid entities on same team, or make bullets not
hit their owners.

Also, the turrets appear to be colliding with the bodies they are
attached to, at least initially. We don't want this to happen.

Turrets need to be made to aim in a particular direction.

Turret positions etc are hard coded at the moment, as I just wanted to
get something working. Needs to be genericised.