Dynamic configuration of dynamic components
Added a "setup()" method to Component taking a **kwargs dict. The
**kwargs from create_game_object() gets passed into this, allowing the
code to configure dynamically created components. Body now looks for a
"position" key, and uses this for an initial position if possible. This
means turrets are now created in the right place.

I think this needs to be rationalised. I think GameObject.initialise()
can be deleted, and Component.setup() perhaps renamed "initialise()".
The various setters on Body should be removed, to prevent it from being
used improperly. I also need to have a look through and unify similar
conceps (as already stated with setup() and initialise(), but also
__init__() methods.) I feel like it can be simplified.