Fix exception fired by physics callback
For whatever reason, the physics callback was resulting in
kill() being called twice on game objects. It could well be
that this double kill()-ing is a bug - but for now, silence
the exceptions by checking is_garbage() in kill() and only
do the business once - probably the right behaviour for kill()
anyway.