Draw using abstract renderer
In order to speed up the drawing I'd like to use OpenGL.  I'd
like to be able to verify & benchmark an OpenGL drawing implementation
against the existing pygame software drawing (and don't want to leave
the game in a non-working state while implementing OpenGL drawing.)

So I've created an abstract renderer interface through which all
drawing will now be done. I've derived a pygame software implementation
into which our existing calls to pygame.draw / Surface.blit etc have
been moved.

The renderer instance is created dynamically based on a type specified
in the config file, so we can add new renderers and plug them in
dynamically.

Internally, the renderer queues a list of jobs for the frame and then
executes them all at once. The intention is that the jobs can be sorted
by layer / rendering state and can then be executed efficiently - but
this isn't done yet, it just processes the list in creation order.