Better background; gamma correct textures
The current background image is crap, use a better one off the internet.
It looks like it's free to use...

Also, since we're applying gamma correction to the resulting image, we
need to apply the reverse mapping to textures or they get gamma
corrected twice, since they're already tuned for display on a monitor!