Start work simplifying drawing
The 'Drawing' object is no longer a component manager. The 'Drawable'
component is no more.  Instead, 'Drawing' is a plain old object that
queries the entity manager for entities with certain combinations of
components and draws them accordingly.

This work isnt finished, in this state the program is broken.

TODO:
  - Remove all references to 'Drawable' in configs
  - Sort out update & state in 'Text' component.
  - Implement 'EntityManager.query()'.
  - Make sure it all works.