Fix physics glitches with turrets
Turrets would often appear in the wrong place and behave very stangely.

The issue was that the telport() function used to position enemies finds
attached entities by querying Joints in the entity manager - but query()
ignores entities yet to be added to the update set, so teleport() wasn't
finding the turrets.

I fixed the problem by adding a query() that will not ignore entities
yet to be created.  However this weirdness with entities yet to be
created might cause problems in future, I might need to find a more
robust solution.