Add docking with planets
Made it possible to 'dock' with planets. This displays a text window
with some detail on the planet.

Actually, you can 'dock' with any 'Dockable' entity, not just planets.
But for now, only planets are dockable.

To do this, celestial bodies were reworked - they are now kinematic
pymunk bodies (instead of being their own special thing) so that
existing spatial query code can be used on them. They are driven by a
system that sets their velocity to maintain the appearance of an orbit.